Added Migration for menu to change primary key of menu in bigint

+++ InventorySystem/InventorySystem.Core/Entities/Menu.cs
@@ -7,7 +7,7 @@ using System.Threading.Tasks;
 
 namespace InventorySystem.Core.Entities
 {
-    public class Menu : AuditableEntity<int>
+    public class Menu : AuditableEntity<long>
     {
         public RootMenu RootMenu{ get; set; }
         public string AreaName { get; set; }
